backend/requirements.txt
# Core web framework & server
fastapi>=0.111.0,<0.112.0
uvicorn>=0.30.1,<0.31.0
websockets>=12.0,<13.0

# Quantum stack
qiskit>=1.3,<2.1
qiskit-ibm-runtime>=0.41.0,<0.42.0
qiskit-optimization>=0.6.1,<0.7.0

# OR-Tools & graph utilities
ortools>=9.10.4067,<9.11.0
osmnx>=1.9.1,<1.10.0
networkx>=3.3,<3.4

# Geo & data
folium>=0.17.0,<0.18.0
geopy==2.4.1
numpy>=1.26.4,<1.27.0
pandas>=2.2.2,<2.3.0

# Flask stack (for app.py compatibility)
Flask>=3.1.2,<3.2.0
Flask-Cors>=6.0.1,<6.1.0
python-dotenv>=1.1.1,<1.2.0

# FastAPI extras
python-multipart>=0.0.9,<0.1.0
pydantic>=2.7.4,<2.8.0
